requirements.txt
munch==4.0.0
numpy==1.24.4
opencv_python==4.9.0.80
Pillow==10.3.0
psutil==5.9.8
streamlit==1.33.0
streamlit_tensorboard==0.0.2
toml==0.10.2
ultralytics==8.1.42
